reset all rule tables before applying rules to score items
the first call to _apply_rules in the render function (i.e. applying
rules to the first score item)  was
setting the status "applied" of each rule to True, causing
pending rules tables for subsequent items to be empty.
By resetting the rule tables (setting applied keyword of the rule dictionary
to False) this problem is solved.