Revert having setter for _View's width
It doesn't make sense for a view's width to be re-settable: in a char the
right, left etc. comes directly from glyph's bbox, what sense could it
possibly have the be able to move the right or left of a char?! If you
need moving a width or right or left, put the view inside of a form.
That's what form is for.