Decrement right padding from non-times if needed
It was ugly to have a right side padding of a non-time object which is
bigger than the space after a note (shortest under notes of a staff).
I wanted to have non-times right padding modified (decremented) if the
above case happens, so that it is assured that notes have got their proper
spaces. Non-times get then just whatever space is left there!