Add rule for centering single notes in bar
I had the problem with adding this functionality when it was directly
added at the end of horizontal spacing rule. Problem was, the
coordinates of the staff content weren't correct, because the lineup
necessary for them hadn't take place yet. The lineup happens first after
the rule is applied (horizontal spacing on staff). This was the reason
for seeing different x cooridnates on Inkscape and in terminal, and the
reason why the functionality put at the end of horizontal spacing was
not
making desired results, since there are horizontal coordinates (e.g. x,
left, right) involved.